+++ platform/mac/framework/AGDnsProxy.h
@@ -94,7 +94,7 @@ typedef void (^logCallback)(AGLogLevel level, const char *msg, int length);
  *      tls://1.1.1.1 -- DNS-over-TLS
  *      https://dns.adguard.com/dns-query -- DNS-over-HTTPS
  *      sdns://... -- DNS stamp (see https://dnscrypt.info/stamps-specifications)
- *      quic://dns.adguard.com:8853 -- DNS-over-QUIC
+ *      quic://dns.adguard.com:853 -- DNS-over-QUIC
  */
 @property(nonatomic, readonly) NSString *address;
 /**