+++ platform/windows/capi/include/ag_dns.h
@@ -40,7 +40,7 @@ typedef struct {
      *     tls://dns.adguard.com -- DNS-over-TLS
      *     https://dns.adguard.com/dns-query -- DNS-over-HTTPS
      *     sdns://... -- DNS stamp (see https://dnscrypt.info/stamps-specifications)
-     *     quic://dns.adguard.com:8853 -- DNS-over-QUIC
+     *     quic://dns.adguard.com:853 -- DNS-over-QUIC
      */
     const char *address;