+++ platform/windows/cs/Adguard.Dns/Adguard.Dns/Api/DnsProxyServer/Configs/UpstreamOptions.cs
@@ -18,7 +18,7 @@ namespace Adguard.Dns.Api.DnsProxyServer.Configs
         /// tls://1.1.1.1 -- DNS-over-TLS
         /// https://dns.adguard.com/dns-query -- DNS-over-HTTPS
         /// sdns://... -- DNS stamp (see https://dnscrypt.info/stamps-specifications)
-        /// quic://dns.adguard.com:8853 -- DNS-over-QUIC
+        /// quic://dns.adguard.com:853 -- DNS-over-QUIC
         /// </summary>
         [ManualMarshalStringToPtr]
         public string Address { get; set; }